Abstract functions out into a helpers.py file and clean up some of the error-catching
