+++ src/khoj/routers/api.py
@@ -5,9 +5,7 @@ import time
 import yaml
 import logging
 import json
-from datetime import datetime
 from typing import List, Optional, Union
-from functools import partial
 
 # External Packages
 from fastapi import APIRouter, HTTPException, Header, Request
@@ -15,8 +13,6 @@ from sentence_transformers import util
 
 # Internal Packages
 from khoj.configure import configure_processor, configure_search
-from khoj.processor.conversation.gpt import converse, extract_questions
-from khoj.processor.conversation.utils import message_to_log, reciprocal_conversation_to_chatml
 from khoj.search_type import image_search, text_search
 from khoj.search_filter.date_filter import DateFilter
 from khoj.search_filter.file_filter import FileFilter
@@ -37,6 +33,10 @@ from khoj.utils.state import SearchType
 from khoj.utils import state, constants
 from khoj.utils.yaml import save_config_to_file_updated_state
 from fastapi.responses import StreamingResponse, Response
+from khoj.routers.helpers import perform_chat_checks, generate_chat_response
+from khoj.processor.conversation.gpt import extract_questions
+from fastapi.requests import Request
+
 
 # Initialize Router
 api = APIRouter()
@@ -417,7 +417,7 @@ def chat_history(
     referer: Optional[str] = Header(None),
     host: Optional[str] = Header(None),
 ):
-    _perform_chat_checks()
+    perform_chat_checks()
 
     # Load Conversation History
     meta_log = state.processor_config.conversation.meta_log
@@ -449,30 +449,23 @@ async def chat(
     referer: Optional[str] = Header(None),
     host: Optional[str] = Header(None),
 ) -> Response:
-    _perform_chat_checks()
-    compiled_references, inferred_queries = await _get_top_k_results(request, q, (n or 5))
+    perform_chat_checks()
+    compiled_references, inferred_queries = await extract_references_and_questions(request, q, (n or 5))
 
-    gpt_response = _generate_chat_response(
-        request,
+    # Get the (streamed) chat response from GPT.
+    gpt_response = generate_chat_response(
         q,
-        client,
-        user_agent,
-        referer,
-        host,
         meta_log=state.processor_config.conversation.meta_log,
         compiled_references=compiled_references,
         inferred_queries=inferred_queries,
     )
-
     if gpt_response is None:
-        return Response(content="No response from GPT", media_type="text/plain", status_code=500)
+        return Response(content=gpt_response, media_type="text/plain", status_code=500)
 
     if stream:
         return StreamingResponse(gpt_response, media_type="text/event-stream", status_code=200)
 
-    if isinstance(gpt_response, str):
-        return Response(content=gpt_response, media_type="text/plain", status_code=200)
-
+    # Get the full response from the generator if the stream is not requested.
     aggregated_gpt_response = ""
     while True:
         try:
@@ -484,21 +477,23 @@ async def chat(
 
     response_obj = {"response": actual_response, "context": compiled_references}
 
+    user_state = {
+        "client_host": request.client.host if request.client else None,
+        "user_agent": user_agent or "unknown",
+        "referer": referer or "unknown",
+        "host": host or "unknown",
+    }
+
+    state.telemetry += [
+        log_telemetry(
+            telemetry_type="api", api="chat", client=client, app_config=state.config.app, properties=user_state
+        )
+    ]
+
     return Response(content=json.dumps(response_obj), media_type="application/json", status_code=200)
 
 
-def _perform_chat_checks():
-    if (
-        state.processor_config is None
-        or state.processor_config.conversation is None
-        or state.processor_config.conversation.openai_api_key is None
-    ):
-        raise HTTPException(
-            status_code=500, detail="Set your OpenAI API key via Khoj settings and restart it to use Khoj Chat."
-        )
-
-
-async def _get_top_k_results(
+async def extract_references_and_questions(
     request: Request,
     q: str,
     n: int,
@@ -527,78 +522,3 @@ async def _get_top_k_results(
             compiled_references = [item.additional["compiled"] for item in result_list]
 
     return compiled_references, inferred_queries
-
-
-def _generate_chat_response(
-    request: Request,
-    q: str,
-    client: Union[str, None],
-    user_agent: Union[str, None],
-    referer: Union[str, None],
-    host: Union[str, None],
-    meta_log: dict,
-    compiled_references: List[str] = [],
-    inferred_queries: List[str] = [],
-):
-    def _save_to_conversation_log(
-        q: str,
-        gpt_response: str,
-        user_message_time: str,
-        compiled_references: List[str],
-        inferred_queries: List[str],
-        meta_log,
-    ):
-        state.processor_config.conversation.chat_session += reciprocal_conversation_to_chatml([q, gpt_response])
-        state.processor_config.conversation.meta_log["chat"] = message_to_log(
-            q,
-            gpt_response,
-            user_message_metadata={"created": user_message_time},
-            khoj_message_metadata={"context": compiled_references, "intent": {"inferred-queries": inferred_queries}},
-            conversation_log=meta_log.get("chat", []),
-        )
-
-    # Load Conversation History
-    meta_log = state.processor_config.conversation.meta_log
-
-    # Initialize Variables
-    api_key = state.processor_config.conversation.openai_api_key
-    chat_model = state.processor_config.conversation.chat_model
-    user_message_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    conversation_type = "general" if q.startswith("@general") else "notes"
-
-    # Switch to general conversation type if no relevant notes found for the given query
-    conversation_type = "notes" if compiled_references else "general"
-    logger.debug(f"Conversation Type: {conversation_type}")
-
-    user_state = {
-        "client_host": request.client.host if request.client else None,
-        "user_agent": user_agent or "unknown",
-        "referer": referer or "unknown",
-        "host": host or "unknown",
-    }
-
-    state.telemetry += [
-        log_telemetry(
-            telemetry_type="api", api="chat", client=client, app_config=state.config.app, properties=user_state
-        )
-    ]
-
-    try:
-        with timer("Generating chat response took", logger):
-            partial_completion = partial(
-                _save_to_conversation_log,
-                q,
-                user_message_time=user_message_time,
-                compiled_references=compiled_references,
-                inferred_queries=inferred_queries,
-                meta_log=meta_log,
-            )
-
-            gpt_response = converse(
-                compiled_references, q, meta_log, model=chat_model, api_key=api_key, completion_func=partial_completion
-            )
-
-    except Exception as e:
-        gpt_response = str(e)
-
-    return gpt_response
